requirements.txt
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.0
polygon-api-client==1.13.4
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
technic==0.0.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
websocket-client==1.7.0
websockets==12.0
